pyproject.toml
[tool.poetry]
name = "luizalabs"
version = "0.1.0"
description = ""
authors = ["Thiago Tancredi Trajano <thiagotancredii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
alembic = "^1.13.2"
pydantic-settings = "^2.5.2"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pyjwt = "^2.9.0"
fastapi-pagination = "^0.12.27"
psycopg = {extras = ["binary"], version = "^3.2.2"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.5"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.init_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev luizalabs/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=luizalabs -vv'
post_test = 'coverage html'